requirements.txt
apturl==0.5.2
beautifulsoup4==4.12.2
blinker==1.4
bs4==0.0.1
catfish==4.16.3
certifi==2019.11.28
chardet==3.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
dnspython==2.3.0
filelock==3.0.12
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
jmespath==1.0.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
menulibre==2.2.2
more-itertools==4.2.0
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
pexpect==4.6.0
psutil==5.5.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
pymacaroons==0.13.0
PyNaCl==1.3.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pyxdg==0.26
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
ssimplejson==3.16.0
six==1.14.0
soupsieve==2.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.8
wadllib==1.3.3
xkit==0.0.0

